+++ include/dca/phys/dca_step/cluster_solver/shared_tools/accumulation/tp/g4_helper.cuh
@@ -130,12 +130,12 @@ inline __device__
 void G4Helper::getComputeRange(const int& my_rank, const int& mpi_size,
                                const uint64_t& total_G4_size, uint64_t & start, uint64_t& end) const {
 
-    unsigned int offset = 0;
+    uint64_t offset = 0;
     // check if originally flattened one-dimensional G4 array can be equally (up to 0) distributed across ranks
     // if balanced, each rank has same amount of elements to compute
     // if not, ranks with (rank_id < nb_more_work_ranks) has to compute 1 more element than other ranks
     bool balanced = (total_G4_size % mpi_size == 0);
-    int local_work = total_G4_size / mpi_size;
+    uint64_t local_work = total_G4_size / mpi_size;
 
     if(balanced) {
         offset = my_rank  * local_work;
